V2.0
added genres appbar and movies selected according to genres list view

+++ popcorn/lib/model/movie.dart
@@ -17,5 +17,5 @@ class Movie {
         backPoster = json["backdrop_path"],
         poster = json["poster_path"],
         overview = json["overview"],
-        rating = json["vote_average"].toDouble();
+        rating = json["vote_average"];
 }
